requirements.txt
mcp-server
python-dotenv
requests
xmltodict
